# app/routers/jobs.py
from typing import Annotated, Any
from uuid import UUID

from fastapi import APIRouter, Depends, HTTPException, status
from psycopg import Connection
from redis import Redis

from app.database import get_connection
from app.dependencies import get_current_user_id
from app.logging import log_event
from app.redis_client import get_redis
from app.repositories import jobs as job_repository
from app.repositories import records as record_repository
from app.schemas.jobs import JobClaimResponse, JobFailureRequest, JobListResponse, JobLockRequest, JobResponse
from app.services import jobs as job_service
from app.services import redis_cache

router = APIRouter(prefix="/jobs", tags=["jobs"])


@router.get("/{job_id}", response_model=JobResponse)
def get_job(
    job_id: UUID,
    connection: Annotated[Connection[dict[str, Any]], Depends(get_connection)],
    redis_client: Annotated[Redis, Depends(get_redis)],
    user_id: Annotated[UUID, Depends(get_current_user_id)],
) -> dict[str, Any]:
    job = job_repository.get_job_for_user(connection, job_id, user_id)
    if job is None:
        log_event("job.not_found", user_id=str(user_id), job_id=str(job_id))
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail="Job not found.",
        )
    return job_service.apply_cached_status(redis_client, job)


@router.get("/records/{record_id}", response_model=JobListResponse)
def list_record_jobs(
    record_id: UUID,
    connection: Annotated[Connection[dict[str, Any]], Depends(get_connection)],
    redis_client: Annotated[Redis, Depends(get_redis)],
    user_id: Annotated[UUID, Depends(get_current_user_id)],
) -> dict[str, list[dict[str, Any]]]:
    record = record_repository.get_record(connection, user_id, record_id)
    if record is None:
        log_event("record.not_found", user_id=str(user_id), record_id=str(record_id))
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail="Record not found.",
        )

    jobs = [
        job_service.apply_cached_status(redis_client, job)
        for job in job_repository.list_jobs_for_record(connection, record_id)
    ]
    return {"jobs": jobs}


@router.post("/claim", response_model=JobClaimResponse)
def claim_job(
    connection: Annotated[Connection[dict[str, Any]], Depends(get_connection)],
    redis_client: Annotated[Redis, Depends(get_redis)],
) -> dict[str, Any]:
    claimed = job_service.claim_next_job(connection, redis_client)
    if claimed is None:
        log_event("job.claim_empty")
        return {"job": None, "lock_token": None}

    job, lock_token = claimed
    connection.commit()
    log_event(
        "job.claimed",
        job_id=str(job["job_id"]),
        record_id=str(job["record_id"]),
        job_type=job["job_type"],
        status=job["status"],
    )
    return {"job": job, "lock_token": lock_token}


@router.post("/{job_id}/succeed", response_model=JobResponse)
def succeed_job(
    job_id: UUID,
    payload: JobLockRequest,
    connection: Annotated[Connection[dict[str, Any]], Depends(get_connection)],
    redis_client: Annotated[Redis, Depends(get_redis)],
) -> dict[str, Any]:
    job = job_repository.mark_succeeded(connection, job_id)
    if job is None:
        log_event("job.succeed_failed", job_id=str(job_id), reason="running_job_not_found")
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail="Running job not found.",
        )

    connection.commit()
    redis_cache.cache_job_state(redis_client, job_id, job["status"])
    redis_cache.release_job_lock(redis_client, job_id, payload.lock_token)
    log_event(
        "job.succeeded",
        job_id=str(job["job_id"]),
        record_id=str(job["record_id"]),
        job_type=job["job_type"],
        status=job["status"],
    )
    return job


@router.post("/{job_id}/fail", response_model=JobResponse)
def fail_job(
    job_id: UUID,
    payload: JobFailureRequest,
    connection: Annotated[Connection[dict[str, Any]], Depends(get_connection)],
    redis_client: Annotated[Redis, Depends(get_redis)],
) -> dict[str, Any]:
    job = job_repository.mark_failed(
        connection=connection,
        job_id=job_id,
        error_code=payload.error_code,
        error_message=payload.error_message,
    )
    if job is None:
        log_event("job.fail_failed", job_id=str(job_id), reason="running_job_not_found")
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail="Running job not found.",
        )

    connection.commit()
    redis_cache.cache_job_state(redis_client, job_id, job["status"])
    redis_cache.release_job_lock(redis_client, job_id, payload.lock_token)
    log_event(
        "job.failed",
        job_id=str(job["job_id"]),
        record_id=str(job["record_id"]),
        job_type=job["job_type"],
        status=job["status"],
        last_error_code=job["last_error_code"],
    )
    return job


@router.post("/{job_id}/cancel", response_model=JobResponse)
def cancel_job(
    job_id: UUID,
    payload: JobLockRequest,
    connection: Annotated[Connection[dict[str, Any]], Depends(get_connection)],
    redis_client: Annotated[Redis, Depends(get_redis)],
) -> dict[str, Any]:
    job = job_repository.cancel_job(connection, job_id)
    if job is None:
        log_event("job.cancel_failed", job_id=str(job_id), reason="active_job_not_found")
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail="Active job not found.",
        )

    connection.commit()
    redis_cache.cache_job_state(redis_client, job_id, job["status"])
    redis_cache.release_job_lock(redis_client, job_id, payload.lock_token)
    log_event(
        "job.canceled",
        job_id=str(job["job_id"]),
        record_id=str(job["record_id"]),
        job_type=job["job_type"],
        status=job["status"],
    )
    return job